+++ src/vec3.rs
@@ -3,27 +3,27 @@ use std::ops::{Add, Div, Mul, Neg, Sub};
 #[derive(Copy, Clone, Debug)]
 #[repr(C)]
 pub struct Vec3 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 impl Vec3 {
-    pub fn new(x: f64, y: f64, z: f64) -> Vec3 {
+    pub fn new(x: f32, y: f32, z: f32) -> Vec3 {
         Vec3 { x, y, z }
     }
     pub fn zero() -> Vec3 {
         Vec3::from_one(0.0)
     }
 
-    pub fn from_one(v: f64) -> Vec3 {
+    pub fn from_one(v: f32) -> Vec3 {
         Vec3 { x: v, y: v, z: v }
     }
 
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         self.norm().sqrt()
     }
 
-    pub fn norm(&self) -> f64 {
+    pub fn norm(&self) -> f32 {
         self.x * self.x + self.y * self.y + self.z * self.z
     }
 
@@ -36,7 +36,7 @@ impl Vec3 {
         }
     }
 
-    pub fn dot(&self, other: &Vec3) -> f64 {
+    pub fn dot(&self, other: &Vec3) -> f32 {
         self.x * other.x + self.y * other.y + self.z * other.z
     }
 
@@ -57,11 +57,11 @@ impl Vec3 {
     }
 
     pub fn near_zero(&self) -> bool {
-        const S: f64 = 1e-8;
+        const S: f32 = 1e-8;
         self.x.abs() < S && self.y.abs() < S && self.z.abs() < S
     }
 
-    pub fn to_vec(&self) -> Vec<f64> {
+    pub fn to_vec(&self) -> Vec<f32> {
         vec![self.x, self.y, self.z]
     }
 
@@ -110,10 +110,10 @@ impl Mul for Vec3 {
     }
 }
 
-impl Mul<f64> for Vec3 {
+impl Mul<f32> for Vec3 {
     type Output = Vec3;
 
-    fn mul(self, other: f64) -> Vec3 {
+    fn mul(self, other: f32) -> Vec3 {
         Vec3 {
             x: self.x * other,
             y: self.y * other,
@@ -122,10 +122,10 @@ impl Mul<f64> for Vec3 {
     }
 }
 
-impl Div<f64> for Vec3 {
+impl Div<f32> for Vec3 {
     type Output = Vec3;
 
-    fn div(self, other: f64) -> Vec3 {
+    fn div(self, other: f32) -> Vec3 {
         Vec3 {
             x: self.x / other,
             y: self.y / other,
@@ -146,7 +146,7 @@ impl Div for Vec3 {
     }
 }
 
-impl Div<Vec3> for f64 {
+impl Div<Vec3> for f32 {
     type Output = Vec3;
 
     fn div(self, other: Vec3) -> Vec3 {
@@ -154,7 +154,7 @@ impl Div<Vec3> for f64 {
     }
 }
 
-impl Mul<Vec3> for f64 {
+impl Mul<Vec3> for f32 {
     type Output = Vec3;
 
     fn mul(self, other: Vec3) -> Vec3 {
